Fixes ctrlc and a verifier database error
The database error was simply that we were sending an empty string
instead of not seeing the env var. Now we don't set if there is no
database to set.

The ctrlc issue is much harder to describe.
When a user hits ctrl+c, a thread is spawned to run our closure, which
attempts to lock on all the known started `container_id`s and stop them.
The issue is that all of these things happen in an async fashion -
starting a container is an http request that returns success
immediately, but in reality takes some time before the `container_id`
returns and is set to be stopped later. So, it was possible to hit ctrl+c
at a specific time where the request was sent, but the id not returned,
and the app would think everything was great and exit, but now there is
an orphaned container that the application started.

The fix to this is to use a poor-man's future (since I could not get
actual futures, from the Rust stdlib, to work in any way) as the type
contained in the threadsafe datastructure. This type, when created, says
"I have no `container_id` and I will never receive one." Later, when we
know *that* we will be starting a container, we tell this type to expect
a `container_id` eventually. When ctrl+c happens now, it polls on this
type and if it is expecting a `container_id`, we poll until it is set by
the main thread, at which point we stop it and move on.

Additionally, we added a number of trips to pause the current thread
after the useful work has been done and before new containers are
started. This is helpful in the verification step, for example, since it
will be starting a verifier one time per test type. The application must
be allowed to continue to eventually set that `container_id`, but it
must trip and stop doing working if ctrl+c was pressed *before* the next
verifier container is started.