Fixes a bug and a QoL issue with verification output
The bug was causing the application to panic and crash in the event that
a container had previous exited when the toolset attempted to stop it.
This usually occurs when an application crashes and Docker cleans up the
container because nothing is running anymore. Now, the toolset fails
more gracefully and continues with its business.

The QoL issue with verifications is an improvement. The original
implementation would attempt to attach to the running verifier container
to stream logs. This worked well when the verifier was still running at
the time `attach_to_container` attempted to attach; however, if the
container had exited (in the case where the verifier had finished its
job successfully) at the time `attach_to_container` was called, then
instead of exiting, it would stream [nothing] forever.

The fix was simply moving that `attach_to_container` call to a separate
thread and blocking the main thread until the container had exited. If
by the time we call `wait_for_container_to_exit` the container had
already exited, then we regain control in the main thread immediately.